+++ src/app/js/app.js
@@ -19,7 +19,6 @@
     ,'orrportal.kw'
     ,'orrportal.v2r'
     ,'orrportal.m2r'
-    ,'orrportal.voc'
     ,'orrportal.firebase'
     ,'orrportal.upload'
     ,'orrportal.util'